compat decoder (WIP)
